mood-analysis-system/requirements.txt
# =============================================================================
# requirements.txt - ALL dependencies in one file
# =============================================================================

# Flask API server
flask==3.0.0
flask-cors==4.0.0

# Audio analysis
librosa==0.10.1
soundfile==0.12.1
praat-parselmouth==0.4.3
numpy==1.24.3

# Text sentiment analysis
vaderSentiment==3.3.2

# LLM providers
openai==1.3.0
google-generativeai==0.3.2
anthropic==0.18.0

# Utilities
python-dotenv==1.0.0
sounddevice  # Optional, for live recording
